# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handler.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  internal_handler.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/internal_handler.APIError'
      meta:
        $ref: '#/definitions/internal_handler.Meta'
      success:
        type: boolean
    type: object
  internal_handler.Meta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  internal_handler.initiatePaymentResponse:
    properties:
      confirmation_url:
        type: string
    type: object
...
  internal_handler.paymentResponse:
    properties:
      amount:
        type: integer
      booking_id:
        type: string
      created_at:
        type: string
      currency:
        type: string
      id:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      provider:
        type: string
      refund_amount:
        type: integer
      refunded_at:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  internal_handler.statusResponse:
    properties:
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: API for the Bani bathhouse marketplace platform.
  title: Bani API
  version: "1.0"
paths:
...
  /bookings/{id}/pay:
    post:
      description: Creates a payment for a booking via YooKassa and returns the confirmation
        URL for redirect
      parameters:
      - description: Booking ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.initiatePaymentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Initiate payment
      tags:
      - payments
  /bookings/{id}/payment:
    get:
      description: Returns the payment associated with a specific booking
      parameters:
      - description: Booking ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.paymentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Get booking payment
      tags:
      - payments
...
  /my/payments:
    get:
      description: Returns a paginated list of payments for the authenticated user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_handler.paymentResponse'
                  type: array
                meta:
                  $ref: '#/definitions/internal_handler.Meta'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: List my payments
      tags:
      - payments
...
  /webhooks/yookassa:
    post:
      consumes:
      - application/json
      description: Processes payment status updates from YooKassa. Returns 200 even
        for domain errors to prevent infinite retries.
      parameters:
      - description: YooKassa webhook payload
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.statusResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      summary: YooKassa webhook
      tags:
      - payments
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"